Refactoring and moving
'Official' QuPath extensions are now in a qupath.ext package.
