+++ src/main/java/qupath/ext/align/gui/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Classes to help with interactive image alignment in QuPath.
  */
-package qupath.lib.gui.align;
+package qupath.ext.align.gui;
