+++ src/main/java/qupath/ext/align/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Experimental commands that are still in development.
  */
-package qupath.experimental;
+package qupath.ext.align;
