+++ src/main/resources/META-INF/services/qupath.lib.gui.extensions.QuPathExtension
@@ -1,1 +1,1 @@
-qupath.align.ExperimentalExtension
+qupath.ext.align.AlignExtension
